+++ lib/typescript/index.d.ts
@@ -1,4 +1,6 @@
 export { AppSprint } from "./AppSprint";
+export { AppSprintAppleAds } from "./AppSprintAppleAds";
 export { NativeAppSprint } from "./NativeAppSprint";
+export type { AppSprintAppleAdsConfig, } from "./AppSprintAppleAds";
 export type { AppSprintConfig, AppSprintOptions, EventType, EventParams, AttributionResult, InstallResponse, DeviceInfo, LogLevel, TestEventResult, NativeAppSprintModule, } from "./types";
 //# sourceMappingURL=index.d.ts.map