+++ src/index.ts
@@ -1,5 +1,9 @@
 export { AppSprint } from "./AppSprint";
+export { AppSprintAppleAds } from "./AppSprintAppleAds";
 export { NativeAppSprint } from "./NativeAppSprint";
+export type {
+  AppSprintAppleAdsConfig,
+} from "./AppSprintAppleAds";
 export type {
   AppSprintConfig,
   AppSprintOptions,